requirements.txt
beautifulsoup4==4.11.1
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.3
idna==3.3
importlib-metadata==4.11.3
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
PyJWT==2.4.0
python-dotenv==0.20.0
pytz==2022.1
requests==2.27.1
soupsieve==2.3.2.post1
twilio==7.9.0
urllib3==1.26.9
Werkzeug==2.1.2
zipp==3.8.0
